replace hatem's components with new login and sinup pages, import them to app.jsx
