+++ frontend/src/App.jsx
@@ -6,14 +6,11 @@ import Header from './components/common/Header';
 import Footer from './components/common/Footer';
 import Homepage from './pages/Homepage';
 import LinkPage from './pages/LinkPage';
+import LoginPage from './pages/LoginPage';
+import SignupPage from './pages/SignupPage';
 import ProfilePage from './pages/ProfilePage';
 import OneCollectionPage from './pages/OneCollectionPage';
-import CollectionPage from './pages/CollectionPage';
 import CollectionsListPage from './pages/CollectionsListPage';
-// import LoginPage from './pages/LoginPage'; // Hatem
-// import SignUpPage from './pages/SignUpPage'; // Hatem
-import LoginPage from './components/authorisation/LoginPage';
-import SignupPage from './components/authorisation/SignupPage';
 import AddLinkForm from './components/collection/AddLinkForm';
 import EditLinkForm from './components/collection/EditLinkForm';
 
@@ -31,12 +28,10 @@ function App() {
             <Routes>
               <Route path='/' element={<Homepage />} />
               <Route path='/link/:linkId' element={<LinkPage />} />
-              <Route path='/signup' element={<SignupPage />} />
-              <Route path='/login' element={<LoginPage />} />
 
               {/* SignUp and LogIn*/}
-              {/* <Route path='/login' element={<LoginPage />} />
-              <Route path='/signup' element={<SignUpPage />} /> */}
+              <Route path='/signup' element={<SignupPage />} />
+              <Route path='/login' element={<LoginPage />} />
 
               {/* Link operations */}
               <Route
