--- frontend/src/components/auth/common/Button.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-const Button = ({onClick, type = 'button', className = '' }) => {
-  return (
-    <button
-      type={type}
-      onClick={onClick}
-      className={`bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline ${className}`}
-    >
-      
-    </button>
-  );
-};
-
-export default Button;
